Complete Ui portion of A-MoreOOP
