+++ src/main/java/MissMinutes.java
@@ -1,4 +1,3 @@
-import java.util.Scanner;
 import java.util.ArrayList;
 import java.io.IOException;
 import java.io.FileInputStream;
@@ -7,17 +6,8 @@ import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 
 public class MissMinutes {
-    private final Scanner stdin;
+    private Ui ui;
     private ArrayList<Task> tasks;
-    private static final String separator = "-".repeat(60);
-    private static final String logo =
-            " __  __ _           __  __ _             _                  \n" +
-            " |  \\/  (_)         |  \\/  (_)           | |              \n" +
-            " | \\  / |_ ___ ___  | \\  / |_ _ __  _   _| |_ ___  ___    \n" +
-            " | |\\/| | / __/ __| | |\\/| | | '_ \\| | | | __/ _ \\/ __| \n" +
-            " | |  | | \\__ \\__ \\ | |  | | | | | | |_| | ||  __/\\__ \\\n" +
-            " |_|  |_|_|___/___/ |_|  |_|_|_| |_|\\__,_|\\__\\___||___/  \n" +
-            "                                                            \n";
 
     public enum CommandType {
         BYE, LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, UNKNOWN
@@ -46,14 +36,8 @@ public class MissMinutes {
     }
 
     public MissMinutes() {
-        this.stdin = new Scanner(System.in);
+        this.ui = new Ui();
         this.tasks = new ArrayList<Task>(100);
-    }
-
-    public void sendMsg(String body) {
-        System.out.println(separator);
-        System.out.println(body);
-        System.out.println(separator);
     }
 
     public Task createTask(String input) throws MissMinutesException {
@@ -80,7 +64,7 @@ public class MissMinutes {
         String reply = "Got it. I've added this task: \n" +
                 task + "\n" +
                 "Now you have " + this.tasks.size() + " tasks in the list.";
-        this.sendMsg(reply);
+        ui.sendMsg(reply);
     }
 
     public void deleteTask(String input) throws MissMinutesException {
@@ -97,35 +81,10 @@ public class MissMinutes {
             String reply = "Noted. I've removed this task:\n"
                     + curr + "\n"
                     + "Now you have " + this.tasks.size() + " tasks in the list.";
-            this.sendMsg(reply);
+            ui.sendMsg(reply);
         } catch (IndexOutOfBoundsException err) {
             throw new MissMinutesException("This task doesn't exist!", err);
         }
-    }
-
-    public void greet() {
-        this.sendMsg("Hello! I'm \n" + logo
-                    + "What can I do for you");
-    }
-
-    public void exit() {
-        this.sendMsg("Bye. Hope to see you again soon!");
-        this.stdin.close();
-    }
-
-    public void printTasks() {
-        if (this.tasks.isEmpty()) {
-            this.sendMsg("There are no tasks in your list.");
-            return;
-        }
-        StringBuilder reply = new StringBuilder("Here are the tasks in your list: ");
-        for (int i = 0; i < this.tasks.size(); i++) {
-            reply.append("\n")
-                    .append((i + 1))
-                    .append(". ")
-                    .append(this.tasks.get(i));
-        }
-        this.sendMsg(reply.toString());
     }
 
     public void markTask(int idx) throws MissMinutesException {
@@ -133,7 +92,7 @@ public class MissMinutes {
             Task curr = this.tasks.get(idx);
             curr.markAsDone();
             String reply = "Nice! I've marked this task as done: \n" + curr;
-            this.sendMsg(reply);
+            ui.sendMsg(reply);
         } catch (IndexOutOfBoundsException err) {
             throw new MissMinutesException("This task doesn't exist!", err);
         }
@@ -144,24 +103,26 @@ public class MissMinutes {
             Task curr = this.tasks.get(idx);
             curr.unmark();
             String reply = "OK, I've marked this task as not done yet: \n" + curr;
-            this.sendMsg(reply);
+            ui.sendMsg(reply);
         } catch (IndexOutOfBoundsException err) {
             throw new MissMinutesException("This task doesn't exist!", err);
         }
     }
 
     public void run() {
+        ui.greet();
         while (true) {
-            String request = this.stdin.nextLine();
+            String request = ui.getInput();
             CommandType cmdType = parseCommand(request);
             boolean tasksChanged = false;
 
             try {
                 switch (cmdType) {
                     case BYE:
+                        ui.exit();
                         return;
                     case LIST:
-                        printTasks();
+                        ui.printTasks(this.tasks);
                         break;
                     case MARK:
                     case UNMARK:
@@ -194,14 +155,14 @@ public class MissMinutes {
                         throw new MissMinutesException("Oh, I'm sowwy, I didn't undewstand dat. (>_<) Can I hewp wif sumthin' else, pwease? UwU");
                 }
             } catch (MissMinutesException err) {
-                this.sendMsg(err.getMessage());
+                ui.sendMsg(err.getMessage());
             }
 
             if (tasksChanged) {
                 try {
                     this.saveTasks();
                 } catch (IOException err) {
-                    this.sendMsg("Failed to save `tasks.bin`: " + err.getMessage());
+                    ui.sendMsg("Failed to save `tasks.bin`: " + err.getMessage());
                 }
             }
         }
@@ -234,13 +195,11 @@ public class MissMinutes {
         try {
             mm.loadTasks();
         } catch (IOException err) {
-            mm.sendMsg("No `tasks.bin` found, creating empty task list");
+            mm.ui.sendMsg("No `tasks.bin` found, creating empty task list");
         } catch (ClassNotFoundException err) {
-            mm.sendMsg(err.getMessage());
+            mm.ui.sendMsg(err.getMessage());
         }
 
-        mm.greet();
         mm.run();
-        mm.exit();
     }
 }
